permission flow

+++ ui/src/app/components/core/drive/DriveCtrl.js
@@ -36,8 +36,9 @@
         })
       }
 
-      $scope.children = function (id) {
+      $scope.children = function (id, access) {
         DriveService.fetch_entity_children(id, user_id).then(function(response) {
+          var access_level = compareAccess(id, 'fr', access)
           data = response.data
           $scope.folders.length = 0
           $scope.files.length = 0
@@ -45,7 +46,7 @@
             if (elem['type'] == 'fr') $scope.folders = elem['items']
             else if (elem['type'] == 'fl') $scope.files = elem['items']
           })
-          AclService.pushContext(AclService.Context.build(id, 'owner'))
+          AclService.pushContext(AclService.Context.build(id, access_level))
           checkIfRootLevel()
         }, function(error) {
             ToastService.showToast(error)
@@ -62,44 +63,61 @@
       }
 
       $scope.add = function (name, e_type) {
-        id = AclService.extractContext().getParentId()
-        DriveService.add_entity(user_id, e_type, name, id).then(function(response) {
-          if (response.status == 200) {
-            $scope.children(id)
-            ToastService.showToast('Success')
-          }
-        }, function(error) {
-            ToastService.showToast(error)
-        })
+        current = AclService.extractContext()
+        current_id = AclService.getParentId(current)
+        current_access = AclService.getAccessLevel(current)
+        if (current_access != 'read') {
+          DriveService.add_entity(user_id, e_type, name, current_id).then(function(response) {
+            if (response.status == 200) {
+              $scope.children(current_id,current_access)
+              ToastService.showToast('Success')
+            }
+          }, function(error) {
+              ToastService.showToast(error)
+          })
+        } else {
+            ToastService.showToast('You don\'t have permission for this action!')
+        }
       }
 
-      $scope.edit = function (id, e_type, new_name) {
-        DriveService.edit_entity(id, user_id, e_type, new_name).then(function(response) {
-          if (response.status == 200) {
-            $scope.children(id)
-            ToastService.showToast('Success')
-          }
-        }, function(error) {
-            ToastService.showToast(error)
-        })
+      $scope.edit = function (id, e_type, new_name, access) {
+        var access_level = compareAccess(id, e_type, access)
+        if (access_level != 'read') {
+          DriveService.edit_entity(id, user_id, e_type, new_name).then(function(response) {
+            if (response.status == 200) {
+              $scope.children(id)
+              ToastService.showToast('Success')
+            }
+          }, function(error) {
+              ToastService.showToast(error)
+          })
+        } else {
+            ToastService.showToast('You don\'t have permission for this action!')
+        }
       }
 
-      $scope.delete = function(id, e_type) {
-        DriveService.remove_entity(id, user_id, e_type).then(function(response) {
-          if(response.status == 200) {
-            $scope.children(id)
-            ToastService.showToast('Success')
-          }
-        }, function(error) {
-            ToastService.showToast(error)
-        })
+      $scope.delete = function(id, e_type, access) {
+        var access_level = compareAccess(id, e_type, access)
+        if (access_level != 'owner') {
+          DriveService.remove_entity(id, user_id, e_type).then(function(response) {
+            if(response.status == 200) {
+              $scope.children(id)
+              ToastService.showToast('Success')
+            }
+          }, function(error) {
+              ToastService.showToast(error)
+          })
+        } else {
+            ToastService.showToast('You don\'t have permission for this action!')
+        }
       }
 
       $scope.navigateBack = function() {
         AclService.popContext()
         current = AclService.extractContext()
         current_id = AclService.getParentId(current)
-        $scope.children(current_id)
+        current_access = AclService.getAccessLevel(current)
+        $scope.children(current_id, current_access)
         if (current_id == 0) $scope.shared(user_id)
         checkIfRootLevel()
       }
@@ -111,6 +129,20 @@
           $scope.showBackNav = false
           $scope.rootLevel = true
         }
+      }
+
+      var compareAccess = function(id, e_type, access) {
+        current = AclService.extractContext()
+        current_access = AclService.getAccessLevel(current)
+        access_level = (typeof access !== 'undefined')? access : current_access
+        var fetch_access = ''
+        DriveService.check_access(id, user_id, e_type).then(function(response) {
+          if(response.data.length) fetch_access = response.data[0][0]
+        }, function(error) {
+          ToastService.showToast(error)
+        })
+        access_level = fetch_access || access_level
+        return access_level
       }
 
       this.init()
